+++ backend/src/main/java/com/pants/backend/repository/ReceiptRepository.java
@@ -2,6 +2,6 @@ package com.pants.backend.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.pants.backend.entity.Receipt;
 
-public interface ReceiptRepository extends JpaRepository<Receipt, Long> {
+public interface ReceiptRepository extends JpaRepository<Receipt, Integer> {
 
 }
